+++ packages/build-mpa-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 4.0.3
+
+- feat: not render tabbar in frm container
+
+## 4.0.2
+
+- fix match rules for app.json
+
 ## 4.0.1
 
 - feat add renderData of `isMPA`