+++ packages/plugin-app-core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.0.3
+
+- [feat] format rax plugin name
+
+## 2.0.2
+
+- [feat] support load `styl` style
+- [fix] ts type of renderComponent
+
 ## 2.0.1
 
 - [feat] set `WEBPACK_CACHE_ID` value