+++ packages/plugin-react-app/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2.0.3
+
+- [feat] support apply static loader with query like `?worker|?url|?raw|?worker&inline`
+- [fix] typo of `vitePlugins`
+- [fix] bump `esbuild` version to `^0.13.12`
+
+## 2.0.2
+
+- [fix] bump version of `webpack-plugin-import`
+
 ## 2.0.1
 
 - [fix] bump version of `@builder/pack`